Windows: Quote command invocation when the &shell is not used
Neovim invokes external commands in two ways, system("cmd arg") calls
cmd through the &shell and applies shell escaping options (just like in Vim).
But system(['cmd', 'arg']) finds cmd in the PATH and executes the command
directly with no shell escaping - both cases are handled by libuv's uv_spawn.

Internaly libuv uses execvp() for UNIX and CreateProcessW() for Windows. The
later takes the arguments as a single string and needs escaping in order to
separate arguments. Libuv does this by default, but we disabled this beaviour
in #3305 to enable system("...") to work as intended, however the previous
behaviour is needed for system([...]).

This commit changes the internal Neovim API (Process) to enable selective
quoting in windows - we want libuv quoting whenever the &shell is not used.
By default quoting is enabled (as was before #3305), but is explicitly set
for system() and jobstart().

Added a functional test for system([...]) vs system("...") verifies the
number of arguments received by an external script is the same for both.

Fixes #3747, unblocks #2773, changed docs and comments to address #3518.