cmake: Add option to disable unittest target
Added a new CMake option ENABLE_UNITTESTS (ON by default) to
disable the 'unittest' target, replacing it with a bogus target
that prints a message about the option.

When luajit/ffi is not available stop with a fatal error, but print
an error message about -DENABLE_UNITTESTS.

The default behaviour remains unchanged, a missing luajit ffi causes
a build to fail, but prints a more informative info message.